feat(api): add api server and reorganise project file structure
