+++ api/.env.example
@@ -6,4 +6,6 @@ MYSQL_PASSWORD=
 MYSQL_DATABASE=
 MYSQL_ROOT_PASSWORD=
 MYSQL_PORT=
-MYSQL_VOLUME=
+
+API_PORT=
+LOG_LEVEL=
