+++ api/database/sql/create-table.js
@@ -1,5 +1,5 @@
-import logger from '../../web/src/util/logger.js';
-import connection from '../sql-connection.js';
+import logger from '../../src/lib/logger.js';
+import connection from './sql-connection.js';
 
 const createToDoTable = 'CREATE TABLE IF NOT EXISTS to_do (id INT AUTO_INCREMENT, userId INT, createdAt TIMESTAMP, updatedAt TIMESTAMP, completeBy TIMESTAMP, title VARCHAR(255), description VARCHAR(255), priority INT, PRIMARY KEY (id))';
