+++ api/database/sql/truncate-table.js
@@ -1,5 +1,5 @@
-import logger from '../../web/src/util/logger.js';
-import connection from '../sql-connection.js';
+import logger from '../../src/lib/logger.js';
+import connection from './sql-connection.js';
 
 const truncateToDoTable = 'TRUNCATE TABLE to_do';
