+++ api/package.json
@@ -3,12 +3,18 @@
   "version": "0.0.0",
   "description": "API for example-app",
   "main": "index.js",
+  "type": "module",
   "dependencies": {
-    "express": "^4.18.2"
+    "dotenv": "^16.4.7",
+    "express": "^4.18.2",
+    "mysql2": "^3.11.3"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "json-server --watch ./db.json"
+    "api": "node ./src/server.js",
+    "createTable": "node ./database/sql/create-table.js",
+    "seedTable": "node ./database/sql/seed-table.js",
+    "truncateTable": "node ./database/sql/truncate-table.js"
   },
   "author": "",
   "license": "MIT",
